[net] Move eviction logic to its own file
